+++ CHANGELOG.md
@@ -2,6 +2,7 @@ CHANGELOG for Sulu
 ==================
 
 * dev-master
+    * HOTFIX      #2864 [MediaBundle]         Return 404 http code for not existing media
     * HOTFIX      #2804 [ListBuilder]         Fixed ids-query in doctrine-list-builder
     * HOTFIX      #2839 [WebsiteBundle]       Include port in URLs